backend/requirements.txt
fastapi
uvicorn
pillow
stegano
python-multipart
